Add utils for redis
