+++ websocket/crates/infra/src/persistence/redis/mod.rs
@@ -7,3 +7,5 @@ pub mod keys;
 pub mod redis_client;
 pub mod types;
 pub mod updates;
+mod utils;
+pub use utils::*;
